JuliaOS/python/pyproject.toml
[project]
name = "juliaos"
version = "0.1.1"
requires-python = ">=3.11"
dependencies = [

  # Dependencies for generated code:

  "urllib3>=1.25.3",
  "python-dateutil>=2.8.2",
  "pydantic>=1.10.5,<2",
  "aenum>=3.1.11",
  
  # Environment variable loading:
  "python-dotenv>=1.1.1"

]

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]